requirements.txt
# Requirements for this Discord bot

discord.py
aiohttp
